Add Generate helper module
Helper module for generating specifications.
